requirements.txt
protobuf==6.31.1
python-dotenv==1.1.0
streamlit==1.43.1
